+++ locales/deDE.lua
@@ -50,7 +50,6 @@ if LOCALE == "deDE" then -- German
 	L["INFO_MSG_LOOT_ENABLED"] 					= "Beute Schnellmodus aktiviert.";
 	L["INFO_MSG_LOOT_DISABLED"] 				= "Beute Schnellmodus deaktiviert.";
 	L["INFO_MSG_MISCELLANEOUS"]					= "Verschiedene Funktionen";
-	L["INFO_MSG_POSTMASTER"]					= "Der Postmeister";
 	L["INFO_MSG_RESULTS"]						= " ergebnis(se)";
 	-- OBJECT TYPES			
 	L["IS_CREATURE"] 							= "Kreatur";