+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Changed
 - [Mark of Honor] is ignored once again.
+- The Postmaster source has been removed.
 
 ### Fixed
 - Output should no longer show in Quiet mode.